pyproject.toml
[project]
name = "django-airport-service"
version = "0.1.0"
description = "Airport API built with Django REST Framework"
authors = [
  { name = "Pavlo Semenikhin", email = "pavel@example.com" }
]
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "Django==5.2.6",
  "djangorestframework==3.16.1",
  "djangorestframework-simplejwt==5.5.1",
  "drf-spectacular==0.28.0",
  "asgiref==3.9.1",
  "attrs==25.3.0",
  "inflection==0.5.1",
  "jsonschema==4.25.1",
  "jsonschema-specifications==2025.9.1",
  "mccabe==0.7.0",
  "pillow==11.3.0",
  "psycopg2-binary==2.9.9",
  "python-dotenv==1.1.1",
  "pyyaml==6.0.2",
  "referencing==0.36.2",
  "rpds-py==0.27.1",
  "sqlparse==0.5.3",
  "typing-extensions==4.15.0",
  "tzdata==2025.2",
  "uritemplate==4.2.0"
]

[project.optional-dependencies]
dev = [
  "coverage==7.10.6",
  "flake8==7.3.0",
  "flake8-quotes==3.4.0",
  "flake8-variables-names==0.0.6"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
